fixed conflicts; separated reply-compose logic; fixed bug in handleClose()
